operating system

+++ 003_chapter3/simplefan.c
@@ -12,7 +12,7 @@ int main(int argc, char *argv[]){
 	for(i=1;i<n;i++)
 		if((childpid=fork())<=0)
 		break;
-	fprintf(stderr, "i:%d process ID:%ld parent ID:%ld child ID:%ld\n",i,(long)getpid(),(long)getppid,(long)childpid);
+	fprintf(stderr, "i:%d process ID:%d parent ID:%d child ID:%d\n",i,getpid(),getppid(),childpid);
 	return 0;
 }
